feature: router, globalStyle 세팅
