+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "axios": "^1.6.2",
     "react": "^18.2.0",
+    "react-cookie": "^6.1.1",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.0",
     "styled-components": "^6.1.1"
   },
   "devDependencies": {
